requirements/test.txt
django
mysql-replication
